aggregate/monthly_upload/requirements.txt
airtable-python-wrapper==0.15.1
google-cloud-secret-manager==2.0.0
